Criado a integração de Orders e Payments com RestTemplate.

+++ ws-rasmoo-plus/src/main/java/com/client/ws/rasmooplus/dto/wsraspay/OrderDto.java
@@ -18,5 +18,5 @@ public class OrderDto {
 
     private String productAcronym;
 
-    private Long discount;
+    private BigDecimal discount;
 }
